refactor: split entity_converter.py and block_item_generator.py into entity/ subpackage
- Created entity/ subpackage with modular structure per Issue #1276:
  - nbt_parser.py: NBT tag and Java property extraction
  - attribute_mapper.py: Entity attribute → Bedrock component mapping
  - block_state_parser.py: Block state variant resolution
  - block_generator.py: Bedrock block behavior JSON assembly
  - item_generator.py: Bedrock item JSON + model resolution
  - type_registry.py: Java entity/block/item type → Bedrock identifier lookups
  - entity_converter.py: Main entity converter implementation
  - block_item_generator.py: Main block/item generator implementation

- Converted entity_converter.py and block_item_generator.py to thin wrappers
  that re-export from the entity/ subpackage, maintaining backward compatibility
- All existing tests pass (66 tests)
- Public API remains unchanged
